+++ tests/functional/test_functional.py
@@ -57,7 +57,7 @@ class TestFunctional(unittest.TestCase):
 
     def test_write_resources_in_region(self):
         """It is sufficient for this not to throw an exception."""
-        self.wanderer.write_resources(regions=["us-east-1"], service_names=["iam"], exclude_resources=[])
+        self.wanderer.write_resources(regions=["us-east-1"], exclude_resources=[])
 
     def test_write_resource_type(self):
         """It is sufficient for this not to throw an exception."""