+++ tests/integration/aws_interface/test_boto3_service.py
@@ -52,19 +52,6 @@ class TestCloudWandererBoto3Service(unittest.TestCase):
         iam_service = self.services.get_service("iam")
         assert iam_service.region == "us-east-1"
 
-    def test_should_query_resources_in_region_regional_service(self):
-        assert self.service.should_query_resources_in_region
-
-    def test_should_query_resources_in_region_global_service_regional_resources(self):
-        assert self.s3_service.should_query_resources_in_region
-
-    def test_should_query_resources_in_region_global_service_regional_resources_wrong_query_region(self):
-        s3_service = self.services.get_service("s3", region_name="eu-west-2")
-        assert not s3_service.should_query_resources_in_region
-
-    def test_should_query_resources_in_region_global_service_global_resources(self):
-        assert self.iam_service.should_query_resources_in_region
-
     def test_account_id(self):
         assert self.service.account_id == "123456789012"
 
@@ -105,3 +92,8 @@ class TestCloudWandererBoto3Service(unittest.TestCase):
             "ap-east-1",
             "eu-west-2",
         ]
+
+    def test__get_collection_from_boto3_resource_type_bad_resource(self):
+        result = self.service._get_collection_from_boto3_resource_type(boto3_resource_type="Test")
+
+        assert result is None