+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added `requiresLoadForFullMetadata` to service mappings. Allows resources like Regional WebACLs to declare that they get impoverished data from their list operation and need `resource.load()` called to fetch all info.
 - Added support for services which have both regional *and* global resources
+- Moved get_and_cleanup_actions from CloudWandererBoto3Resource to ResourceMap
+- Added Resource Filtering options to flexibly filter out resources you're not interested in.
 
 ## New Resources
  - Added ec2 nat gateways